Update logQoS behavior
Previously MediaSession.logQos() behaved statelessly, in a sense that the MediaQoS object provided as the argument, would be identical to the one observed in the generated MediaEvent. This behavior has been updated so that MediaSession retains an internal MediaQoS object and will now merge the provided MediaQoS argument with the internal object. The result of this behavior is that if you do not provide a value for a MediaQoS field that you have previously provided within the same MediaSession, the most recent value that was provided for that field will be merged into the object.